Refactor OpenClaw Rust helpers to satisfy structural rules
