+++ crates/clawdentity-cli/src/commands/connector.rs
@@ -5,34 +5,31 @@ use std::time::Duration;
 use anyhow::{Result, anyhow};
 use clap::Subcommand;
 use clawdentity_core::config::ConfigPathOptions;
-use clawdentity_core::db::now_utc_ms;
-use clawdentity_core::db_inbound::{
-    InboundPendingItem, append_inbound_event, delete_pending, list_pending_due,
-    mark_pending_attempt, move_pending_to_dead_letter, upsert_pending,
-};
-use clawdentity_core::http::client as create_http_client;
 use clawdentity_core::runtime_openclaw::OpenclawRuntimeConfig;
 use clawdentity_core::{
-    CONNECTOR_FRAME_VERSION, ConnectorClient, ConnectorClientOptions, ConnectorClientSender,
-    ConnectorFrame, ConnectorServiceInstallInput, ConnectorServiceUninstallInput, CoreError,
-    DeliverAckFrame, DeliverFrame, RuntimeServerState, SqliteStore, flush_outbound_queue_to_relay,
-    install_connector_service, new_frame_id, now_iso, spawn_connector_client,
+    ConnectorClient, ConnectorClientOptions, ConnectorClientSender, ConnectorServiceInstallInput,
+    ConnectorServiceUninstallInput, CoreError, RuntimeServerState, SqliteStore,
+    flush_outbound_queue_to_relay, install_connector_service, spawn_connector_client,
     uninstall_connector_service,
 };
-use serde_json::{Value, json};
+use serde_json::json;
 use tokio::sync::watch;
 use tokio::task::JoinHandle;
 
 const DEFAULT_CONNECTOR_PORT: u16 = 19400;
 const DEFAULT_OPENCLAW_HOOK_PATH: &str = "/hooks/wake";
-const CONNECTOR_RETRY_DELAY_MS: i64 = 5_000;
-const INBOUND_RETRY_INTERVAL: Duration = Duration::from_secs(1);
-const INBOUND_RETRY_BATCH_SIZE: usize = 50;
-const INBOUND_MAX_ATTEMPTS: i64 = 3;
 const OUTBOUND_FLUSH_INTERVAL: Duration = Duration::from_millis(500);
 const OUTBOUND_FLUSH_BATCH_SIZE: usize = 50;
 
+mod delivery;
 mod runtime_config;
+
+use delivery::{run_inbound_loop, run_inbound_retry_loop};
+
+#[cfg(test)]
+use delivery::{
+    build_deliver_ack_reason, build_openclaw_hook_payload, should_dead_letter_after_failure,
+};
 
 #[derive(Debug, Subcommand)]
 pub enum ConnectorCommand {
@@ -361,61 +358,6 @@ fn spawn_inbound_retry_task(
     tokio::spawn(async move { run_inbound_retry_loop(store, openclaw_runtime, shutdown_rx).await })
 }
 
-async fn run_inbound_loop(
-    mut connector_client: ConnectorClient,
-    relay_sender: ConnectorClientSender,
-    store: SqliteStore,
-    openclaw_runtime: OpenclawRuntimeConfig,
-    mut shutdown_rx: watch::Receiver<bool>,
-) -> Result<()> {
-    let hook_url = openclaw_runtime.hook_url()?;
-    let http_client = create_http_client()?;
-
-    loop {
-        tokio::select! {
-            _ = shutdown_rx.changed() => {
-                if *shutdown_rx.borrow() {
-                    return Ok(());
-                }
-            }
-            frame = connector_client.recv_frame() => {
-                let Some(frame) = frame else {
-                    return Ok(());
-                };
-                match frame {
-                    ConnectorFrame::Deliver(deliver) => {
-                        handle_deliver_frame(
-                            &store,
-                            &relay_sender,
-                            &http_client,
-                            &hook_url,
-                            &openclaw_runtime,
-                            deliver,
-                        )
-                        .await;
-                    }
-                    ConnectorFrame::EnqueueAck(ack) => {
-                        if ack.accepted {
-                            tracing::debug!(
-                                ack_id = %ack.ack_id,
-                                "relay accepted outbound enqueue frame"
-                            );
-                        } else {
-                            let reason = ack.reason.as_deref().unwrap_or("unknown");
-                            tracing::warn!(
-                                ack_id = %ack.ack_id,
-                                reason,
-                                "relay rejected outbound enqueue frame"
-                            );
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-}
-
 async fn run_outbound_flush_loop(
     store: SqliteStore,
     relay_sender: ConnectorClientSender,
@@ -448,454 +390,6 @@ async fn run_outbound_flush_loop(
             }
         }
     }
-}
-
-async fn run_inbound_retry_loop(
-    store: SqliteStore,
-    openclaw_runtime: OpenclawRuntimeConfig,
-    mut shutdown_rx: watch::Receiver<bool>,
-) -> Result<()> {
-    let hook_url = openclaw_runtime.hook_url()?;
-    let http_client = create_http_client()?;
-    let mut interval = tokio::time::interval(INBOUND_RETRY_INTERVAL);
-    interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-
-    loop {
-        tokio::select! {
-            _ = shutdown_rx.changed() => {
-                if *shutdown_rx.borrow() {
-                    return Ok(());
-                }
-            }
-            _ = interval.tick() => {
-                retry_due_inbound_deliveries(
-                    &store,
-                    &http_client,
-                    &hook_url,
-                    &openclaw_runtime,
-                )
-                .await;
-            }
-        }
-    }
-}
-
-async fn retry_due_inbound_deliveries(
-    store: &SqliteStore,
-    http_client: &reqwest::Client,
-    hook_url: &str,
-    openclaw_runtime: &OpenclawRuntimeConfig,
-) {
-    let due_items = match list_pending_due(store, now_utc_ms(), INBOUND_RETRY_BATCH_SIZE) {
-        Ok(items) => items,
-        Err(error) => {
-            tracing::warn!(error = %error, "failed to list pending inbound deliveries");
-            return;
-        }
-    };
-
-    for item in due_items {
-        retry_pending_inbound_delivery(store, http_client, hook_url, openclaw_runtime, item).await;
-    }
-}
-
-async fn retry_pending_inbound_delivery(
-    store: &SqliteStore,
-    http_client: &reqwest::Client,
-    hook_url: &str,
-    openclaw_runtime: &OpenclawRuntimeConfig,
-    item: InboundPendingItem,
-) {
-    let payload = match serde_json::from_str::<Value>(&item.payload_json) {
-        Ok(payload) => payload,
-        Err(error) => {
-            let reason = format!("invalid pending payload_json: {error}");
-            if let Err(move_error) = move_pending_to_dead_letter(store, &item.request_id, &reason) {
-                tracing::warn!(
-                    error = %move_error,
-                    request_id = %item.request_id,
-                    "failed to move invalid pending payload to dead letter"
-                );
-            }
-            return;
-        }
-    };
-
-    let deliver = DeliverFrame {
-        v: CONNECTOR_FRAME_VERSION,
-        id: item.request_id.clone(),
-        ts: now_iso(),
-        from_agent_did: item.from_agent_did.clone(),
-        to_agent_did: item.to_agent_did.clone(),
-        payload,
-        content_type: Some("application/json".to_string()),
-        conversation_id: item.conversation_id.clone(),
-        reply_to: item.reply_to.clone(),
-    };
-
-    match forward_deliver_to_openclaw(http_client, hook_url, openclaw_runtime, &deliver).await {
-        Ok(()) => {
-            if let Err(error) = delete_pending(store, &item.request_id) {
-                tracing::warn!(
-                    error = %error,
-                    request_id = %item.request_id,
-                    "failed to clear resolved pending inbound delivery"
-                );
-                return;
-            }
-            if let Err(error) = append_inbound_event(
-                store,
-                "delivered_retry",
-                Some(item.request_id.clone()),
-                Some(
-                    json!({
-                        "frameId": item.frame_id,
-                    })
-                    .to_string(),
-                ),
-            ) {
-                tracing::warn!(
-                    error = %error,
-                    request_id = %item.request_id,
-                    "failed to append delivered_retry inbound event"
-                );
-            }
-        }
-        Err(error) => {
-            handle_pending_retry_failure(store, &item, &error);
-        }
-    }
-}
-
-fn handle_pending_retry_failure(
-    store: &SqliteStore,
-    item: &InboundPendingItem,
-    error: &anyhow::Error,
-) {
-    if should_dead_letter_after_failure(item.attempt_count) {
-        let reason = format!("max retry attempts exceeded: {error}");
-        if let Err(move_error) = move_pending_to_dead_letter(store, &item.request_id, &reason) {
-            tracing::warn!(
-                error = %move_error,
-                request_id = %item.request_id,
-                "failed to move pending inbound delivery to dead letter"
-            );
-        }
-        return;
-    }
-
-    let next_attempt_at_ms = now_utc_ms() + CONNECTOR_RETRY_DELAY_MS;
-    if let Err(mark_error) = mark_pending_attempt(
-        store,
-        &item.request_id,
-        next_attempt_at_ms,
-        Some(error.to_string()),
-    ) {
-        tracing::warn!(
-            error = %mark_error,
-            request_id = %item.request_id,
-            "failed to update pending inbound retry attempt"
-        );
-        return;
-    }
-    if let Err(event_error) = append_inbound_event(
-        store,
-        "pending_retry",
-        Some(item.request_id.clone()),
-        Some(
-            json!({
-                "frameId": item.frame_id,
-                "attemptCount": item.attempt_count + 1,
-                "nextAttemptAtMs": next_attempt_at_ms,
-            })
-            .to_string(),
-        ),
-    ) {
-        tracing::warn!(
-            error = %event_error,
-            request_id = %item.request_id,
-            "failed to append pending_retry inbound event"
-        );
-    }
-}
-
-fn should_dead_letter_after_failure(current_attempt_count: i64) -> bool {
-    current_attempt_count.saturating_add(1) >= INBOUND_MAX_ATTEMPTS
-}
-
-async fn handle_deliver_frame(
-    store: &SqliteStore,
-    relay_sender: &ConnectorClientSender,
-    http_client: &reqwest::Client,
-    hook_url: &str,
-    openclaw_runtime: &OpenclawRuntimeConfig,
-    deliver: DeliverFrame,
-) {
-    let delivery_result =
-        forward_deliver_to_openclaw(http_client, hook_url, openclaw_runtime, &deliver).await;
-
-    let persistence_result =
-        persist_inbound_delivery_result(store, &deliver, delivery_result.as_ref()).await;
-    if let Err(error) = persistence_result.as_ref() {
-        tracing::warn!(error = %error, request_id = %deliver.id, "failed to persist inbound delivery result");
-    }
-
-    let ack_reason = build_deliver_ack_reason(
-        delivery_result.as_ref().err(),
-        persistence_result.as_ref().err(),
-    );
-    let ack_accepted = ack_reason.is_none();
-    if let Err(error) = send_deliver_ack(relay_sender, &deliver.id, ack_accepted, ack_reason).await
-    {
-        tracing::warn!(error = %error, request_id = %deliver.id, "failed to send deliver ack");
-    }
-
-    if let Err(error) = delivery_result {
-        tracing::warn!(error = %error, request_id = %deliver.id, to_agent_did = %deliver.to_agent_did, "failed to forward inbound payload to OpenClaw hook");
-    }
-}
-
-fn build_deliver_ack_reason(
-    delivery_error: Option<&anyhow::Error>,
-    persistence_error: Option<&anyhow::Error>,
-) -> Option<String> {
-    match (delivery_error, persistence_error) {
-        (Some(delivery_error), Some(persistence_error)) => Some(format!(
-            "{delivery_error}; failed to persist inbound delivery result: {persistence_error}"
-        )),
-        _ => None,
-    }
-}
-
-async fn forward_deliver_to_openclaw(
-    http_client: &reqwest::Client,
-    hook_url: &str,
-    openclaw_runtime: &OpenclawRuntimeConfig,
-    deliver: &DeliverFrame,
-) -> Result<()> {
-    let mut request = http_client
-        .post(hook_url)
-        .header("content-type", "application/json")
-        .header("x-clawdentity-agent-did", &deliver.from_agent_did)
-        .header("x-clawdentity-to-agent-did", &deliver.to_agent_did)
-        .json(&build_openclaw_hook_payload(
-            &openclaw_runtime.hook_path,
-            deliver,
-        ));
-
-    if let Some(token) = openclaw_runtime
-        .hook_token
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        request = request.header("x-openclaw-token", token);
-    }
-
-    let response = request
-        .send()
-        .await
-        .map_err(|error| anyhow!("openclaw hook request failed: {error}"))?;
-
-    if !response.status().is_success() {
-        return Err(anyhow!("openclaw hook returned HTTP {}", response.status()));
-    }
-
-    Ok(())
-}
-
-fn build_openclaw_hook_payload(hook_path: &str, deliver: &DeliverFrame) -> Value {
-    if normalize_hook_path(hook_path) == "/hooks/wake" {
-        return build_openclaw_wake_payload(deliver);
-    }
-
-    build_openclaw_agent_payload(deliver)
-}
-
-fn build_openclaw_agent_payload(deliver: &DeliverFrame) -> Value {
-    let message = extract_content(&deliver.payload);
-    json!({
-        "message": message,
-        "content": message,
-        "senderDid": deliver.from_agent_did,
-        "recipientDid": deliver.to_agent_did,
-        "requestId": deliver.id,
-        "metadata": {
-            "conversationId": deliver.conversation_id,
-            "replyTo": deliver.reply_to,
-            "payload": deliver.payload,
-        },
-    })
-}
-
-fn build_openclaw_wake_payload(deliver: &DeliverFrame) -> Value {
-    let wake_text = render_openclaw_wake_text(deliver);
-    let session_id = deliver
-        .payload
-        .get("sessionId")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty());
-    let mut payload = json!({
-        "message": wake_text,
-        "text": wake_text,
-        "mode": "now",
-    });
-    if let Some(session_id) = session_id {
-        payload["sessionId"] = Value::String(session_id.to_string());
-    }
-    payload
-}
-
-fn render_openclaw_wake_text(deliver: &DeliverFrame) -> String {
-    let message = extract_content(&deliver.payload);
-    let mut lines = vec![format!(
-        "Clawdentity peer message from {}",
-        deliver.from_agent_did
-    )];
-
-    if !message.trim().is_empty() {
-        lines.push(String::new());
-        lines.push(message);
-    }
-
-    if let Some(request_id) = Some(deliver.id.as_str()).filter(|value| !value.trim().is_empty()) {
-        lines.push(String::new());
-        lines.push(format!("Request ID: {request_id}"));
-    }
-    if let Some(conversation_id) = deliver
-        .conversation_id
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        lines.push(format!("Conversation ID: {conversation_id}"));
-    }
-    if let Some(reply_to) = deliver
-        .reply_to
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        lines.push(format!("Reply To: {reply_to}"));
-    }
-
-    lines.join("\n")
-}
-
-fn extract_content(payload: &Value) -> String {
-    if let Some(content) = payload.get("content").and_then(Value::as_str) {
-        return content.to_string();
-    }
-    if let Some(message) = payload.get("message").and_then(Value::as_str) {
-        return message.to_string();
-    }
-    if let Some(text) = payload.get("text").and_then(Value::as_str) {
-        return text.to_string();
-    }
-    if let Some(text) = payload.as_str() {
-        return text.to_string();
-    }
-    payload.to_string()
-}
-
-#[allow(clippy::too_many_lines)]
-async fn persist_inbound_delivery_result(
-    store: &SqliteStore,
-    deliver: &DeliverFrame,
-    delivery_result: std::result::Result<&(), &anyhow::Error>,
-) -> Result<()> {
-    let received_at_ms = now_utc_ms();
-    let payload_json = deliver.payload.to_string();
-
-    append_inbound_event(
-        store,
-        "received",
-        Some(deliver.id.clone()),
-        Some(
-            json!({
-                "frameId": deliver.id,
-                "fromAgentDid": deliver.from_agent_did,
-                "toAgentDid": deliver.to_agent_did,
-            })
-            .to_string(),
-        ),
-    )?;
-
-    if delivery_result.is_ok() {
-        if let Err(error) = delete_pending(store, &deliver.id) {
-            tracing::warn!(
-                error = %error,
-                request_id = %deliver.id,
-                "failed to clear pending inbound record after successful delivery"
-            );
-        }
-        append_inbound_event(
-            store,
-            "delivered",
-            Some(deliver.id.clone()),
-            Some(json!({ "frameId": deliver.id }).to_string()),
-        )?;
-        return Ok(());
-    }
-
-    let last_error = delivery_result
-        .err()
-        .map(ToString::to_string)
-        .unwrap_or_else(|| "delivery failed".to_string());
-
-    upsert_pending(
-        store,
-        InboundPendingItem {
-            request_id: deliver.id.clone(),
-            frame_id: deliver.id.clone(),
-            from_agent_did: deliver.from_agent_did.clone(),
-            to_agent_did: deliver.to_agent_did.clone(),
-            payload_json,
-            payload_bytes: i64::try_from(deliver.payload.to_string().len()).unwrap_or(i64::MAX),
-            received_at_ms,
-            next_attempt_at_ms: received_at_ms + CONNECTOR_RETRY_DELAY_MS,
-            attempt_count: 1,
-            last_error: Some(last_error),
-            last_attempt_at_ms: Some(received_at_ms),
-            conversation_id: deliver.conversation_id.clone(),
-            reply_to: deliver.reply_to.clone(),
-        },
-    )?;
-
-    append_inbound_event(
-        store,
-        "pending",
-        Some(deliver.id.clone()),
-        Some(
-            json!({
-                "frameId": deliver.id,
-                "nextAttemptAtMs": received_at_ms + CONNECTOR_RETRY_DELAY_MS,
-            })
-            .to_string(),
-        ),
-    )?;
-
-    Ok(())
-}
-
-async fn send_deliver_ack(
-    relay_sender: &ConnectorClientSender,
-    ack_id: &str,
-    accepted: bool,
-    reason: Option<String>,
-) -> Result<()> {
-    relay_sender
-        .send_frame(ConnectorFrame::DeliverAck(DeliverAckFrame {
-            v: CONNECTOR_FRAME_VERSION,
-            id: new_frame_id(),
-            ts: now_iso(),
-            ack_id: ack_id.to_string(),
-            accepted,
-            reason,
-        }))
-        .await
-        .map_err(anyhow::Error::from)
 }
 
 pub(super) fn env_trimmed(name: &str) -> Option<String> {
